+++ src/Facades/Currency.php
@@ -7,7 +7,9 @@ use Illuminate\Support\Facades\Facade;
 
 /**
  * @method static array get(\Statamic\Sites\Site $site)
- * @method static string parse($price, \Statamic\Sites\Site $site)
+ * @method static string parse($amount, \Statamic\Sites\Site $site)
+ * @method static int toPence(float $amount)
+ * @method static float toDecimal(int $amount)
  */
 class Currency extends Facade
 {