+++ src/Support/Currency.php
@@ -24,10 +24,22 @@ class Currency implements Contract
             ->first();
     }
 
-    public function parse($price, Site $site): string
+    public function parse($amount, Site $site): string
     {
+        if (is_string($amount)) {
+            if (str_contains($amount, '.')) {
+                $amount = str_replace('.', '', $amount);
+            }
+
+            $amount = (int) $amount;
+        }
+
+        if (is_float($amount)) {
+            $amount = $amount * 100;
+        }
+
         try {
-            $money = new Money(str_replace('.', '', $price), new MoneyCurrency($this->get($site)['code']));
+            $money = new Money(str_replace('.', '', (int) $amount), new MoneyCurrency($this->get($site)['code']));
 
             $numberFormatter = new NumberFormatter($site->locale(), \NumberFormatter::CURRENCY);
             $moneyFormatter = new IntlMoneyFormatter($numberFormatter, new ISOCurrencies());
@@ -36,6 +48,16 @@ class Currency implements Contract
         } catch (\ErrorException $e) {
             throw new CurrencyFormatterNotWorking("Extension PHP-intl not installed.");
         }
+    }
+
+    public function toPence(float $amount): int
+    {
+        return $amount * 100;
+    }
+
+    public function toDecimal(int $amount): float
+    {
+        return $amount / 100;
     }
 
     public static function bindings(): array